Add DeviceRegistrationTests
